+++ History.md
@@ -1,4 +1,14 @@
+0.0.3 / 2017-02-23
+==================
+
+* feat: add convertStringToArray and convertStringToObject
+
+0.0.2 / 2017-02-22
+==================
+
+* fix: update validateProps, add code and error
+
 0.0.1 / 2017-02-21
 ==================
 
-* init project
+* feat: init project